+++ FSW/main.cpp
@@ -15,11 +15,11 @@
 
 // NS2 headers
 #include "src/headers/config.hpp"
-#include "src/headers/comUtil.hpp"
+#include "src/headers/commandHandling.hpp"
 #include "src/headers/eventUtil.hpp"
 #include "src/headers/timingClass.hpp"
 #include "src/headers/timing.hpp"
-#include "src/headers/memUtil.hpp"
+#include "src/headers/dataCollection.hpp"
 
 /* - - - - - - Functions - - - - - - */
